build: adopt hardening warnings + -Werror for project code
adapted from c-template, applied to our sources only -- the lua
submodule stays vanilla behind a quiet static_library. flags are
has_argument-gated: -Wextra -Werror, null-dereference, array-bounds,
format=2, shadow, implicit-fallthrough, strict/missing-prototypes,
missing-declarations, vla, alloca, shift-overflow,
frame-larger-than=4096, and -fsanitize-trap=all (ub traps to ud2
instead of silently continuing; the whole test suite runs clean
under it). -Wfloat-equal was tried and dropped: a libm/strtod
implementation compares floats exactly on purpose.

new src/platform.h gives every platform/arch function a prototype
(-Wmissing-prototypes). real catches from the sweep: a signedness
slip in vsnprintf's %d path, a null-deref-if-child-vanished in
api_spawn, an accidental nested comment, and a header include that
had been spliced into the middle of a block comment.

docs (CODING_STYLE etc) deferred on purpose; flags enforce more than
prose anyway.

+++ src/libc/vsnprintf.c
@@ -404,7 +404,7 @@ vsnprintf(char *buf, size_t n, const char *fmt, va_list ap)
 				v = va_arg(ap, long);
 			else
 				v = va_arg(ap, int);
-			fmt_int(&o, v < 0 ? -(unsigned long long)v : v,
+			fmt_int(&o, v < 0 ? -(unsigned long long)v : (unsigned long long)v,
 			    v < 0, 10, false, prec, width, flags);
 			break;
 		}
